% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.issvd.R
\name{cv.issvd}
\alias{cv.issvd}
\title{Performs cross-validation method to select tuning parameters in the ISSVD.}
\usage{
cv.issvd(x, ncomp, varnumv, varnumu, type = "soft", nfolds)
}
\arguments{
\item{x}{list with K Data matrices of dimension $n x pk$.
We are interested in finding sparse singular value decomposition of
dimension $pk$.}

\item{ncomp}{The number of factors; default is 1.}

\item{varnumv}{list with optional user-supplied varnumv sequence.}

\item{varnumu}{list with optional user-supplied varnumu sequence.}

\item{type}{Which kind of penalty is chosen? Possible options are \code{"soft"}, \code{"hard"}, \code{"bic"},
and \code{"SCAD"}. Default is \code{"soft"}.}

\item{nfolds}{The number of folds - default is 10. Although nfolds can be as large as the sample size (leave-one-out CV), it is not recommended for large datasets.
Smallest value allowable is nfolds=3.}
}
\value{
\item{varv_opt}{Optimal parameters varnumv for v.}
        \item{varu_opt}{Optimal parameter varnumu for u.}
}
\description{
Performs cross-validation method to select tuning parameters in the ISSVD.
}
\examples{
   n=100;p1=p2=p3=50;sig1=sig2=sig3=0.1;
   d1=d2=1;d3=50;
   simdata=simdata(n,p1,p2,p3,sig1,
                 sig2,sig3,d1,d2,d3)
   varnumv=list(c(15,15,15),c(16,16,16),
              c(17,17,17))
   varnumu=list(24,25,26)
   issvd_model1=cv.issvd(x,1,varnumv,
                         varnumu,type='soft',10)
   varv_opt=issvd_model1$varv_opt
   varu_opt=issvd_model1$varu_opt
   issvd_model=issvd(x,1,varv_opt,varu_opt,type='soft')
}
