% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issvd.R
\name{issvd}
\alias{issvd}
\title{Perform intergative and sparse singular value decomposition analysis}
\usage{
issvd(
  x,
  ncomp = 1,
  varnumv,
  varnumu,
  type = "soft",
  maxiter = 100,
  eps = 10^(-4)
)
}
\arguments{
\item{x}{list with K Data matrices of dimension $n x pk$, which can contain NA for missing
values. We are interested in finding sparse singular value decomposition of
dimension $pk$.}

\item{ncomp}{The number of factors in the ISSVD to be returned; default is 1.}

\item{varnumv}{vectors with K elements. How sparse do you want vk to be in the K data matrices?
Each elements of varnumv represents how many variables is left, hence its value must be between 1 and pk.
The smaller it is, the sparser v will be.}

\item{varnumu}{How sparse do you want u to be in the K data matrices?}

\item{type}{Which kind of penalty is chosen? Possible options are \code{"soft"}, \code{"hard"}, \code{"bic"},
and \code{"SCAD"}. Default is \code{"soft"}.}

\item{maxiter}{How many iterations should be performed. It is best to run at
least 100 of so. Default is 100.}

\item{eps}{convergence threshhold. The default threshhold is 10^(-4).}
}
\value{
\item{U}{The common left singular matrix U.}
        \item{D}{The individual scale matrices D.}
        \item{V}{The individual right singular matrices V.}
}
\description{
Performs intergative and sparse singular value decomposition analysis 
to multiple data matrix with penalty on the columns and the rows.
}
\examples{
   n=100;p1=p2=p3=50;sig1=sig2=sig3=0.1;
   d1=d2=1;d3=50;
   simdata=simdata(n,p1,p2,p3,sig1,
                  sig2,sig3,d1,d2,d3)
   varv_opt=list(c(16,16,16))
   varu_opt=25
   issvd_model=issvd(x,1,varv_opt,varu_opt,type='soft')
}
