% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{simdata}
\alias{simdata}
\title{Generate mupltiple-sources simulation datasets.}
\usage{
simdata(n, p1, p2, p3, sig1, sig2, sig3, d1, d2, d3)
}
\arguments{
\item{n}{The number of x.}

\item{ncomp}{The number of factors in the simulated data.}

\item{p}{p1 p2 p3 represent the number of variables in three datasets, respectively.}

\item{sig}{sig1 sig2 sig3 represent the noise ratio in three datasets, respectively.}

\item{d}{d1 d2 d3 represent the scale parameters in three datasets, respectively.}
}
\value{
\item{x}{multi-source data.}
        \item{truex}{multi-source data without noise.}
}
\description{
Generate mupltiple-sources simulation datasets following the settings.
}
